requirements.txt
ultralytics>=8.3.0
torch>=2.1.0
torchvision>=0.16.0
numpy
Pillow
opencv-python-headless
matplotlib
seaborn
scikit-learn
pyyaml
requests
tqdm
pandas
